Split package view code modules to smaller files
Split the package related Django views code classes to multiple smaller
files, roughly one per view class or grouped by the same functionality.

This commit does not include any functionality changes outside of
changes to import locations.
