+++ django/thunderstore/repository/views/__init__.py
@@ -1,25 +1,15 @@
-from .repository import (
-    PackageCreateOldView,
-    PackageCreateView,
-    PackageDetailView,
-    PackageDocsView,
-    PackageDownloadView,
+from .docs import PackageDocsView
+from .package.create import PackageCreateOldView, PackageCreateView
+from .package.detail import PackageDetailView
+from .package.download import PackageDownloadView
+from .package.list import (
     PackageListByDependencyView,
     PackageListByOwnerView,
     PackageListSearchView,
     PackageListView,
-    PackageVersionDetailView,
 )
+from .package.version import PackageVersionDetailView
 
 __all__ = [
-    "PackageCreateOldView",
-    "PackageCreateView",
-    "PackageDetailView",
     "PackageDocsView",
-    "PackageDownloadView",
-    "PackageListByDependencyView",
-    "PackageListByOwnerView",
-    "PackageListSearchView",
-    "PackageListView",
-    "PackageVersionDetailView",
 ]
