+++ django/thunderstore/repository/views/docs.py
@@ -1,0 +1,7 @@
+from django.views.generic import TemplateView
+
+from thunderstore.repository.mixins import CommunityMixin
+
+
+class PackageDocsView(CommunityMixin, TemplateView):
+    template_name = "repository/package_docs.html"
